fix: quota keeper refactoring + casts explained

+++ contracts/credit/CreditConfiguratorV3.sol
@@ -353,7 +353,7 @@ contract CreditConfiguratorV3 is ICreditConfiguratorV3, ACLNonReentrantTrait {
         uint16 liquidationDiscount = PERCENTAGE_FACTOR - liquidationPremium;
         uint16 liquidationDiscountExpired = PERCENTAGE_FACTOR - liquidationPremiumExpired;
 
-        uint16 newLTUnderlying = uint16(liquidationDiscount - feeLiquidation); // I:[CC-18]
+        uint16 newLTUnderlying = liquidationDiscount - feeLiquidation; // I:[CC-18]
         (, uint16 ltUnderlying) = CreditManagerV3(creditManager).collateralTokenByMask(UNDERLYING_TOKEN_MASK);
 
         if (newLTUnderlying != ltUnderlying) {
